CleanUp and Refactoring Project structure of record (view)
