+++ EiDiA_frontend/src/components/record/Dashboard.js
