+++ EiDiA_frontend/src/components/record/RecordSymbol.js
