+++ EiDiA_frontend/src/components/record/Widgets/GraphsWidget.js
