+++ EiDiA_frontend/src/components/record/Widgets/IndicatorWidget.js
