+++ EiDiA_frontend/src/components/record/Widgets/LogWidget.js
