+++ EiDiA_frontend/src/components/record/Widgets/WidgetDropTarget.js
